refactor: move templates components, views, and store to features
