+++ packages/frontend/editor-ui/src/composables/commandBar/useTemplateCommands.ts
@@ -1,6 +1,6 @@
 import { computed } from 'vue';
 import { useI18n } from '@n8n/i18n';
-import { useTemplatesStore } from '@/stores/templates.store';
+import { useTemplatesStore } from '@/features/templates/templates.store';
 import { useCanvasOperations } from '@/composables/useCanvasOperations';
 import type { CommandGroup } from './types';
