+++ packages/frontend/editor-ui/src/experiments/templateRecoV2/stores/templateRecoV2.store.ts
@@ -2,7 +2,7 @@ import { useTelemetry } from '@/composables/useTelemetry';
 import { TEMPLATE_RECO_V2, VIEWS } from '@/constants';
 import { useCloudPlanStore } from '@/stores/cloudPlan.store';
 import { usePostHog } from '@/stores/posthog.store';
-import { useTemplatesStore } from '@/stores/templates.store';
+import { useTemplatesStore } from '@/features/templates/templates.store';
 import { STORES } from '@n8n/stores';
 import { defineStore } from 'pinia';
 import { computed } from 'vue';
