+++ packages/frontend/editor-ui/src/features/templates/components/TemplateDetails.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import TemplateDetailsBlock from '@/components/TemplateDetailsBlock.vue';
+import TemplateDetailsBlock from './TemplateDetailsBlock.vue';
 import NodeIcon from '@/components/NodeIcon.vue';
 import { filterTemplateNodes } from '@/utils/nodeTypesUtils';
 import { abbreviateNumber } from '@/utils/typesUtils';
@@ -10,7 +10,7 @@ import type {
 	ITemplatesWorkflow,
 } from '@n8n/rest-api-client/api/templates';
 import type { ITag } from '@n8n/rest-api-client/api/tags';
-import { useTemplatesStore } from '@/stores/templates.store';
+import { useTemplatesStore } from '@/features/templates/templates.store';
 import TimeAgo from '@/components/TimeAgo.vue';
 import { isFullTemplatesCollection, isTemplatesWorkflow } from '@/utils/templates/typeGuards';
 import { useRouter } from 'vue-router';
