+++ packages/frontend/editor-ui/src/features/templates/components/TemplateDetailsBlock.vue
