+++ packages/frontend/editor-ui/src/features/templates/components/TemplateFilters.vue
