+++ packages/frontend/editor-ui/src/features/templates/components/TemplateList.vue
