+++ packages/frontend/editor-ui/src/features/templates/components/TemplatesInfoCard.vue
