+++ packages/frontend/editor-ui/src/features/templates/templates.store.ts
@@ -17,9 +17,9 @@ import { STORES } from '@n8n/stores';
 import { useRootStore } from '@n8n/stores/useRootStore';
 import { defineStore } from 'pinia';
 import { computed, ref } from 'vue';
-import { useSettingsStore } from './settings.store';
-import { useUsersStore } from './users.store';
-import { useWorkflowsStore } from './workflows.store';
+import { useSettingsStore } from '@/stores/settings.store';
+import { useUsersStore } from '@/stores/users.store';
+import { useWorkflowsStore } from '@/stores/workflows.store';
 
 export interface ITemplateState {
 	categories: ITemplatesCategory[];
