+++ packages/frontend/editor-ui/src/features/templates/views/TemplatesCollectionView.vue
@@ -1,11 +1,11 @@
 <script setup lang="ts">
 import { computed, onMounted, ref, watch } from 'vue';
-import TemplateDetails from '@/components/TemplateDetails.vue';
-import TemplateList from '@/components/TemplateList.vue';
+import TemplateDetails from '../components/TemplateDetails.vue';
+import TemplateList from '../components/TemplateList.vue';
 import TemplatesView from './TemplatesView.vue';
 import type { ITemplatesWorkflow } from '@n8n/rest-api-client/api/templates';
 import { VIEWS } from '@/constants';
-import { useTemplatesStore } from '@/stores/templates.store';
+import { useTemplatesStore } from '@/features/templates/templates.store';
 import { useTemplateWorkflow } from '@/utils/templates/templateActions';
 import { useExternalHooks } from '@/composables/useExternalHooks';
 import { useNodeTypesStore } from '@/stores/nodeTypes.store';
