+++ packages/frontend/editor-ui/src/features/templates/views/TemplatesView.vue
